+++ packages/components-react/src/pages/TextFieldWrapper.tsx
@@ -212,15 +212,33 @@ export const TextFieldWrapperPage = (): JSX.Element => {
         </PTextFieldWrapper>
       </div>
 
-      <div className="playground light" title="should render with unit">
-        <PTextFieldWrapper label="Label with unit" unit="km/h">
+      <div className="playground light" title="should render input type text with unit">
+        <PTextFieldWrapper label="Label with unit input type text" unit="km/h">
+          <input type="text" defaultValue="three hundred" />
+        </PTextFieldWrapper>
+      </div>
+
+      <div className="playground light" title="should render input type number with unit">
+        <PTextFieldWrapper label="Label with unit input type number" unit="km/h">
           <input type="number" defaultValue={300} />
         </PTextFieldWrapper>
       </div>
 
-      <div className="playground light" title="should render with unit position suffix">
-        <PTextFieldWrapper label="Label with unit position suffix" unit="kWh" unitPosition="suffix">
+      <div className="playground light" title="should render input type text with unit position suffix">
+        <PTextFieldWrapper label="Label with unit input type text position suffix" unit="kWh" unitPosition="suffix">
+          <input type="text" defaultValue="four hundred" />
+        </PTextFieldWrapper>
+      </div>
+
+      <div className="playground light" title="should render input type number with unit position suffix">
+        <PTextFieldWrapper label="Label with unit input type number position suffix" unit="kWh" unitPosition="suffix">
           <input type="number" defaultValue={400} />
+        </PTextFieldWrapper>
+      </div>
+
+      <div className="playground light" title="should render counter when counter and unit are set">
+        <PTextFieldWrapper label="Label with counter and unit" unit="km/h">
+          <input type="text" defaultValue="three hundred" maxLength={50} />
         </PTextFieldWrapper>
       </div>
     </>