+++ packages/components/src/components/form/text-field-wrapper/text-field-wrapper-utils.ts
@@ -6,7 +6,10 @@ export type TextFieldWrapperUnitPosition = typeof UNIT_POSITIONS[number];
 
 export const hasCounter = (el: HTMLTextAreaElement | HTMLInputElement): boolean => el.maxLength >= 0;
 export const hasCounterAndIsTypeText = (el: HTMLInputElement): boolean => el.type === 'text' && hasCounter(el);
-export const hasUnitAndIsTypeNumber = (el: HTMLInputElement, unit: string): boolean => !!unit && el.type === 'number';
+export const hasUnitAndIsTypeTextOrNumber = (el: HTMLInputElement, unit: string): boolean => {
+  const { type } = el;
+  return !!unit && (type === 'text' || type === 'number');
+};
 export const setCounterInnerHtml = (el: HTMLTextAreaElement | HTMLInputElement, counterElement: HTMLElement): void => {
   counterElement.innerText = `${el.value.length}/${el.maxLength}`;
 };